+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version 0.2.45
+
+### New feature
+-   Handle PF2e Exploit Vulnerability effects
+-   Z - not trigger move reactions
+
 ## Version 0.2.44
 
 ### New feature